Rename to `precompiledGenesis.json`
